Updated example quasi http application to use Kabomu.Mediator framework.
