+++ examples/Memory.FileExchange/FileServer.cs
@@ -33,7 +33,7 @@ namespace Memory.FileExchange
                 Transport = transport,
                 ErrorHandler = errorHandler,
             };
-            _instance.Application = new FileReceiver2(endpoint, uploadDirPath);
+            _instance.Application = new FileReceiver(endpoint, uploadDirPath);
 
             await hub.AddServer(endpoint, _instance);
